Menu Nodes don't have leaf parameter
It is fix for the possiblyAddExpander() function in OOJSPlus.Item.
One of the conditions for creating an expander is that the node must not be a leaf.
This parameter was previously not passed in the constructor.

ERM-Comment: MenuEditor uses the same tree as BSInsertCategories.

ERM31612

[4.2.x]

Change-Id: If70f95a098279a605ae7090927a3da29d49489fc
(cherry picked from commit bc6c24cbdd0b1e698d2d48d1b52fac69a78f0792)